[DD13] Отдельная загрузка ресурсов для каждой карты

+++ src/GlobalConstants.h
@@ -6,7 +6,7 @@ namespace GlobalConstants
 {
     const string Paused = "paused";
     const string Resumed = "resumed";
-    const int damageTypeCount = 4;
-    const int resourceTypeCount  = 3;
+    const size_t damageTypeCount = 4;
+    const size_t resourceTypeCount  = 3;
 }
